Add support for atrix osh

+++ killrecovery.sh
@@ -1,21 +1,23 @@
 #!/sbin/sh
+
 mkdir -p /sd-ext
-rm /cache/recovery/command
-rm /cache/update.zip
+rm -f /cache/recovery/command
+rm -f /cache/update.zip
 touch /tmp/.ignorebootmessage
-kill $(ps | grep /sbin/adbd)
-kill $(ps | grep /sbin/recovery)
+
+if [ ! -d /system/bootmenu ]; then
+    # kill $(ps | grep /sbin/adbd)
+    kill $(ps | grep /sbin/recovery)
+fi
 
 # On the Galaxy S, the recovery comes test signed, but the
 # recovery is not automatically restarted.
-if [ -f /init.smdkc110.rc ]
-then
+if [ -f /init.smdkc110.rc ]; then
     /sbin/recovery &
 fi
 
 # Droid X
-if [ -f /init.mapphone_cdma.rc ]
-then
+if [ -f /init.mapphone_cdma.rc ]; then
     /sbin/recovery &
 fi
